+++ lib/fsm2.dart
@@ -1,12 +1,12 @@
 export 'src/builders/co_region_builder.dart';
 export 'src/builders/state_builder.dart';
 export 'src/definitions/state_definition.dart'
-    hide onEnter, onExit, addCoRegion;
+    hide addCoRegion, onEnter, onExit;
 export 'src/exceptions.dart';
-export 'src/state_machine.dart';
-export 'src/state_of_mind.dart' hide removePath, addPath;
+export 'src/state_machine.dart' hide findStateDefinition, oldestAncestor;
+export 'src/state_of_mind.dart' hide addPath, removePath;
 export 'src/transitions/transition_definition.dart';
-export 'src/types.dart' show State, Event, OnEnter, OnExit;
+export 'src/types.dart' show Event, OnEnter, OnExit, State;
 export 'src/visualise/smcat_file.dart';
 export 'src/visualise/smcat_folder.dart';
 export 'src/visualise/svg_file.dart';